Initial WIP for adding meta tables for File Download Logs and API Request Logs
#7285
