+++ includes/component-functions.php
@@ -220,6 +220,7 @@ function edd_setup_components() {
 	edd_register_component( 'log_api_request', array(
 		'schema' => '\\EDD\\Database\\Schema\\Logs_Api_Requests',
 		'table'  => '\\EDD\\Database\\Tables\\Logs_Api_Requests',
+		'meta'   => '\\EDD\\Database\\Tables\\Log_Api_Request_Meta',
 		'query'  => '\\EDD\\Database\\Queries\\Log_Api_Request',
 		'object' => '\\EDD\\Logs\\Api_Request_Log',
 		'meta'   => false
@@ -229,6 +230,7 @@ function edd_setup_components() {
 	edd_register_component( 'log_file_download', array(
 		'schema' => '\\EDD\\Database\\Schema\\Logs_File_Downloads',
 		'table'  => '\\EDD\\Database\\Tables\\Logs_File_Downloads',
+		'meta'   => '\\EDD\\Database\\Tables\\Log_File_Download_Meta',
 		'query'  => '\\EDD\\Database\\Queries\\Log_File_Download',
 		'object' => '\\EDD\\Logs\\File_Download_Log',
 		'meta'   => false
